finish infection comparison and little refractor
