+++ src/dashapp/App.py
@@ -1,8 +1,6 @@
-from dash import Dash, html
+from dash import Dash
 import dash
-import dash_bootstrap_components as dbc
 from pages.public.data import *
-from pages.public.css import *
 
 markdown_text = open('data/home/markdown_files/introduction.txt', encoding='utf-8').read()
