+++ src/dashapp/pages/infection.py
@@ -8,6 +8,7 @@ from .public.css import *
 from .public.utils import *
 import dash_bootstrap_components as dbc
 from .public.File_Factory import ModelOne
+import datetime
 
 style_radio = {
     'margin-left': '10px'
