+++ src/dashapp/pages/public/utils_list/evacuation_util.py
@@ -1,0 +1,2 @@
+def find_single_agent_path():
+    pass
